requirements.txt
asgiref==3.8.1
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
PyJWT==2.8.0
python-dateutil==2.8.2
qrcode==7.4.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
